# src/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A股全自动交易系统后端API - 用户注册登录、策略管理、模拟交易、实盘交易、WebSocket实时推送",
        "title": "ZeroQuant 2.0 Backend API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/account": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模拟交易"
                ],
                "summary": "查询模拟账户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Uncle0206061_zeroquant2_backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/account/simulate/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模拟交易"
                ],
                "summary": "创建模拟账户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Uncle0206061_zeroquant2_backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/order/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模拟交易"
                ],
                "summary": "订单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Uncle0206061_zeroquant2_backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/order/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模拟交易"
                ],
                "summary": "提交订单",
                "parameters": [
                    {
                        "description": "订单参数: stock_code, direction(1买入/2卖出), quantity, price(0市价)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Uncle0206061_zeroquant2_backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/order/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模拟交易"
                ],
                "summary": "订单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Uncle0206061_zeroquant2_backend_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模拟交易"
                ],
                "summary": "撤单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Uncle0206061_zeroquant2_backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/position": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模拟交易"
                ],
                "summary": "持仓列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "股票代码",
                        "name": "stock_code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Uncle0206061_zeroquant2_backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/position/{stock_code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模拟交易"
                ],
                "summary": "持仓详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "股票代码",
                        "name": "stock_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Uncle0206061_zeroquant2_backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Uncle0206061_zeroquant2_backend_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码：0=成功，非0=失败",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}